% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.history.R
\name{make.history}
\alias{make.history}
\title{Make History from historical stock data}
\usage{
make.history(quotesdf, windowwidth = 20)
}
\arguments{
\item{quotesdf:}{a stock quotes dataframe with columns Date, Open, and Adj.Close}

\item{windowwidth:}{window size for moving average/standard dev.}
}
\description{
Make History from historical stock data
}
\examples{
calcbuyval( mkbsdf(1,2),maxhold=1 )
}

