% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcbuyval.R
\name{mkbsdf}
\alias{mkbsdf}
\title{make buy sell data frame}
\usage{
mkbsdf(buy, adjclose, safedate = T)
}
\arguments{
\item{buy:}{vector of buy values (expect NA or value)}

\item{adjclose:}{values at close of day}

\item{safedate:}{can the date be used? Default: vector of True}
}
\description{
make a dataframe suitable for calcbuyval
}
\examples{
calcbuyval( mkbsdf(1,2),maxhold=1 )
}

