% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/safedates.R
\name{safedates}
\alias{safedates}
\title{safedates: are dates continous}
\usage{
safedates(d1, d2)
}
\arguments{
\item{d1:}{new date to test if safe}

\item{d2:}{prev date that should be followed by d1 if d1 is safe}
}
\description{
function to check dates are continious (before adding a buy signal)
that is, only at most a holiday separates them
first input should be newer date, the date we will determine is safe or not
-- it's okay if we have data on a holiday. weird, but okay (maybe not NYSE tradded stock)
}
\examples{
safedates(as.Date('2015-01-03'),as.Date('2015-01-02') )
}

